requirements.txt
aniso8601==7.0.0
argon2==0.1.10
attrs==19.1.0
autopep8==1.4.4
boto3==1.9.213
botocore==1.12.213
cached-property==1.5.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.4
django-allauth==0.39.1
django-anymail==6.1.0
django-appconf==1.0.3
django-cors-headers==3.1.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.1
django-graphql-extensions==0.0.6
django-graphql-jwt==0.2.2
django-heroku==0.3.1
django-imagekit==4.0.2
django-model-utils==3.2.0
django-redis==4.10.0
django-resized==0.3.9
django-sendgrid-v5==0.8.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.2
docutils==0.15.2
environ==1.0
flake8==3.7.8
geographiclib==1.49
geopy==1.20.0
googlemaps==3.0.2
graphene==2.1.8
graphene-django==2.5.0
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.1
msgpack-python==0.5.6
oauthlib==3.1.0
packaging==19.1
pexpect==4.7.0
pg==0.1
pilkit==2.0
Pillow==6.1.0
promise==2.2.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
psycopg2-binary==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
PyOpenGL==3.1.0
pytest-django==3.5.1
pytest-sugar==0.9.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-http-client==3.1.0
python-slugify==3.0.3
python3-openid==3.1.0
pytz==2019.2
raven==6.10.0
redis==2.10.6
requests==2.22.0
Rx==3.0.1
s3transfer==0.2.1
sendgrid==6.0.5
sendmail==2.0
sentry-sdk==0.11.1
singledispatch==3.4.0.3
six==1.12.0
slacker==0.13.0
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
twilio==6.29.3
urllib3==1.25.3
whitenoise==4.1.3
